7/2 voting generally works, the corrrect way
voting generally works, the corrrect way.

each event has a different id on the votes, so can call it specifically.

the 1 problem is that it jumps to the top of the page on the trending page when it votes, but that's a problem for another day (aka probably after lunch lol)