7/6 this is kinda working
kinda working.
but u can toggle between 4 numbers. that's not ok. i need to seperate upvoting and switching back to normal. so like that project we did in 247. with the 4 transformations. this can either be regular upvote, regular downvote, or back to normal+regUpVote or  back to normal+regDownVote.

but wanted to committ before i screw everyting up.

or but also - the gray isnt working on the trending events page - it changes the correct event number, but the incorrect event color (always changes event #1 color